requirements.txt
Flask==2.3.2
pymongo==4.6.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0
itsdangerous==2.1.2
python-jose==3.3.0
PyJWT==2.8.0
bcrypt==4.1.2
pandas==2.2.2
matplotlib==3.8.4
flask-dance
Flask-DotEnv
oauthlib
google-auth
google-auth-oauthlib
